Extract HmGui into a separate file/module
