+++ engine/lib/phx/src/ui/hmgui/group.rs
@@ -1,4 +1,4 @@
-use glam::Vec2;
+use glam::{Vec2, Vec4};
 
 use super::*;
